refactor(auth): reorganize biometric files into dedicated directory
Move all biometric-related files into Sources/Auth/Biometrics/ and
extract AuthClient biometric extension into its own file for better
organization.
